+++ packages/v4/patternfly-docs/content/design-guidelines/styles/colors-for-charts/colors-grid.js
@@ -4,7 +4,7 @@ import { Grid, GridItem } from '@patternfly/react-core';
 const ColorsGrid = () => (
   <Grid>
     <GridItem sm={12} md={6} lg={4}>
-      <h3 className="pf-c-title pf-m-xl ws-heading ws-title ws-h3">Blue family</h3>
+      <h3 className="pf-v5-c-title pf-m-xl ws-heading ws-title ws-h3">Blue family</h3>
       <Grid>
         <GridItem span={2}>
           <div className="ws-content-chart-colors ws-content-chart-colors-gallery--blue-1"></div>
@@ -54,7 +54,7 @@ const ColorsGrid = () => (
       </Grid>
     </GridItem>
     <GridItem sm={12} md={6} lg={4}>
-      <h3 className="pf-c-title pf-m-xl ws-heading ws-title ws-h3">Green family</h3>
+      <h3 className="pf-v5-c-title pf-m-xl ws-heading ws-title ws-h3">Green family</h3>
       <Grid>
         <GridItem span={2}>
           <div className="ws-content-chart-colors ws-content-chart-colors-gallery--green-1"></div>
@@ -104,7 +104,7 @@ const ColorsGrid = () => (
       </Grid>
     </GridItem>
     <GridItem sm={12} md={6} lg={4}>
-      <h3 className="pf-c-title pf-m-xl ws-heading ws-title ws-h3">Cyan family</h3>
+      <h3 className="pf-v5-c-title pf-m-xl ws-heading ws-title ws-h3">Cyan family</h3>
       <Grid>
         <GridItem span={2}>
           <div className="ws-content-chart-colors ws-content-chart-colors-gallery--cyan-1"></div>
@@ -154,7 +154,7 @@ const ColorsGrid = () => (
       </Grid>
     </GridItem>
     <GridItem sm={12} md={6} lg={4}>
-      <h3 className="pf-c-title pf-m-xl ws-heading ws-title ws-h3">Purple family</h3>
+      <h3 className="pf-v5-c-title pf-m-xl ws-heading ws-title ws-h3">Purple family</h3>
       <Grid>
         <GridItem span={2}>
           <div className="ws-content-chart-colors ws-content-chart-colors-gallery--purple-1"></div>
@@ -204,7 +204,7 @@ const ColorsGrid = () => (
       </Grid>
     </GridItem>
     <GridItem sm={12} md={6} lg={4}>
-      <h3 className="pf-c-title pf-m-xl ws-heading ws-title ws-h3">Gold family</h3>
+      <h3 className="pf-v5-c-title pf-m-xl ws-heading ws-title ws-h3">Gold family</h3>
       <Grid>
         <GridItem span={2}>
           <div className="ws-content-chart-colors ws-content-chart-colors-gallery--gold-1"></div>
@@ -254,7 +254,7 @@ const ColorsGrid = () => (
       </Grid>
     </GridItem>
     <GridItem sm={12} md={6} lg={4}>
-      <h3 className="pf-c-title pf-m-xl ws-heading ws-title ws-h3">Orange family</h3>
+      <h3 className="pf-v5-c-title pf-m-xl ws-heading ws-title ws-h3">Orange family</h3>
       <Grid>
         <GridItem span={2}>
           <div className="ws-content-chart-colors ws-content-chart-colors-gallery--orange-1"></div>
@@ -304,7 +304,7 @@ const ColorsGrid = () => (
       </Grid>
     </GridItem>
     <GridItem sm={12} md={6} lg={4}>
-      <h3 className="pf-c-title pf-m-xl ws-heading ws-title ws-h3">Red family</h3>
+      <h3 className="pf-v5-c-title pf-m-xl ws-heading ws-title ws-h3">Red family</h3>
       <Grid>
         <GridItem span={2}>
           <div className="ws-content-chart-colors ws-content-chart-colors-gallery--red-1"></div>
@@ -354,7 +354,7 @@ const ColorsGrid = () => (
       </Grid>
     </GridItem>
     <GridItem sm={12} md={6} lg={4}>
-      <h3 className="pf-c-title pf-m-xl ws-heading ws-title ws-h3">Grey family</h3>
+      <h3 className="pf-v5-c-title pf-m-xl ws-heading ws-title ws-h3">Grey family</h3>
       <Grid>
         <GridItem span={2}>
           <div className="ws-content-chart-colors ws-content-chart-colors-gallery--black-1"></div>